:sparkles: IMPROVE: framework detection
