+++ e2e/svelte/package.json
@@ -4,15 +4,17 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"web": "vite --port 3006",
+		"web": "vite --port 3016",
 		"api": "cross-env TZ=utc e2e-api",
 		"dev": "concurrently \"pnpm run web\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
 		"build": "vite build",
 		"tests": "playwright test ",
-		"preview": "vite preview --port 3006",
+		"previewWeb": "vite preview --port 3006",
+		"preview": "concurrently \"pnpm run previewWeb\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
 		"check": "svelte-check --tsconfig ./tsconfig.json"
 	},
 	"devDependencies": {
+		"@kitql/helper": "^0.5.0",
 		"@playwright/test": "1.30.0",
 		"@sveltejs/vite-plugin-svelte": "^2.0.2",
 		"@tsconfig/svelte": "^3.0.0",
