+++ e2e/sveltekit/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "web": "vite dev --host 0.0.0.0",
+    "web": "vite dev --host 0.0.0.0 --port 3017",
     "api": "cross-env TZ=utc e2e-api",
     "dev": "concurrently \"pnpm run web\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
     "build": "vite build",
