+++ packages/houdini/src/cmd/init.ts
@@ -4,7 +4,7 @@ import fetch from 'node-fetch'
 import { execSync } from 'node:child_process'
 import prompts from 'prompts'
 
-import { fs, parseJSON, path, pullSchema } from '../lib'
+import { detectTools, fs, parseJSON, path, pullSchema } from '../lib'
 import type { ConfigFile } from '../runtime/lib/config'
 
 // the init command is responsible for scaffolding a few files
@@ -35,7 +35,7 @@ export default async function init(
 		}, {})
 	}
 
-	// if no path was given, we	'll use cwd
+	// if no path was given, we'll use cwd
 	const targetPath = _path ? path.resolve(_path) : process.cwd()
 
 	// git check
@@ -538,66 +538,6 @@ async function gitIgnore(targetPath: string) {
 
 	if (!existing.includes('\n$houdini\n')) {
 		await fs.writeFile(filepath, existing + '\n$houdini\n')
-	}
-}
-
-type DetectedTools = {
-	typescript: boolean
-	framework: 'kit' | 'sapper' | 'svelte'
-	module: 'esm' | 'commonjs'
-	package_manager: 'npm' | 'yarn' | 'pnpm'
-}
-
-async function detectTools(cwd: string): Promise<DetectedTools> {
-	// if there's no package.json then there's nothing we can detect
-	try {
-		const packageJSONFile = await fs.readFile(path.join(cwd, 'package.json'))
-		if (packageJSONFile) {
-			var packageJSON = JSON.parse(packageJSONFile)
-		} else {
-			throw new Error('not found')
-		}
-	} catch {
-		throw new Error(
-			'❌ houdini init must target an existing node project (with a package.json)'
-		)
-	}
-
-	// grab the dev dependencies
-	const { devDependencies, dependencies } = packageJSON
-
-	const hasDependency = (dep: string) => Boolean(devDependencies?.[dep] || dependencies?.[dep])
-
-	let framework: 'svelte' | 'kit' = 'svelte'
-	if (hasDependency('@sveltejs/kit')) {
-		framework = 'kit'
-	}
-
-	let typescript = false
-	try {
-		await fs.stat(path.join(cwd, 'tsconfig.json'))
-		typescript = true
-	} catch {}
-
-	// package manager?
-	let package_manager: 'npm' | 'yarn' | 'pnpm' = 'npm'
-	let dir = cwd
-	do {
-		if (fs.existsSync(path.join(dir, 'pnpm-lock.yaml'))) {
-			package_manager = 'pnpm'
-			break
-		}
-		if (fs.existsSync(path.join(dir, 'yarn.lock'))) {
-			package_manager = 'yarn'
-			break
-		}
-	} while (dir !== (dir = path.dirname(dir)))
-
-	return {
-		typescript,
-		framework,
-		module: packageJSON['type'] === 'module' ? 'esm' : 'commonjs',
-		package_manager,
 	}
 }
