+++ packages/houdini/src/lib/index.ts
@@ -11,6 +11,7 @@ export * from './introspection'
 export * from './cleanupFiles'
 export * from './deepMerge'
 export * from './plugin'
+export * from './detectTools'
 
 export * from './walk'
 export type { EmbeddedGraphqlDocument } from './walk'
